Extract the Avatar module (avatar fetch) into Api/Avatar
Fourth module of the #777 folder migration. Moves avatar resolution/fetch into
Jellyfin.Plugin.SSO_Auth.Api.Avatar:

- AvatarService, AvatarUrlValidator, AvatarContentType -> Api/Avatar/
  (namespace Api.Avatar), import added at each call site.

Avatar is the first NON-leaf module: AvatarUrlValidator validates fetch targets
through the Net SSRF classifier, so Avatar is allowed to depend on Net and only
Net. Generalises the module-boundary rule from a pure-leaf check into
ApiModule_ImportsOnlyItsAllowedApiModules(module, params allowed): a leaf lists
no allowed modules, Avatar lists Net. The still-unmigrated flat Api core and
non-Api namespaces (Config) stay importable during the migration; a module
never imports its own namespace.

Also drops the cross-module doc cref in the Net IpAddressClassifier (it named
AvatarUrlValidator via <see cref>, which after this move would have coupled the
Net leaf to Avatar) to a plain-text mention.

Pure mechanical move, no behavior change. Verified by the full build
(--warnaserror, both TFMs), the complete test suite (1575 both TFMs), every
existing conformance rule, and the ABI-floor build.

Refs #777
